pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.32",
    "datasets>=3.5.0",
    "huggingface-hub[cli]>=0.30.1",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.11.3",
    "vllm==0.8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
run-batch-job = "llm.__main__:main"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
